+++ site/.vitepress/theme/index.ts
@@ -1,0 +1,3 @@
+import DefaultTheme from "vitepress/theme-without-fonts";
+import "./style.css";
+export default DefaultTheme;